More identity validation and filtering
